fix: refactored useKey.ts
